feat: Add monitoring service
Add monitoring service to the application. This service will be
responsible for monitoring the status of the application and
reporting it to the user. The service is implemented in
the monitoring module. The service is started in the main
function of the application.

This required for the application to have a general
monitoring service that can be used to monitor the status.

Breaking changes: None

Resolves: None