feat: Implements TLS conf for the HTTP monitor
Implements TLS configuration for the HTTP monitor. This includes the
ability to specify the client certificate, server certificate, and the
server's private key. The monitor will use these to establish a secure
connection to the server.

This is required for the monitor to be able to connect to servers that
require client certificates for authentication.

Breaking changes: None

Resolves: None