feat: Add logging configuration file
Add logging framework configuration file to allow for more flexible
logging configuration. This will allow for more flexible logging
configuration without needing to recompile the application. An example
logging configuration file is provided in the resources/test directory.

Breaking changes: Breaks the existing logging configuration and
non daemon will now require loggingfile to be specified when running
the application. stderr is currently not used and will be removed in
a future release.

Resolves: None