feat: Add server module
Implements a server module to handle incoming requests to get the
status of the monitors. The server module uses warp to create a web
server that listens on a specified port and returns the status of the
monitors when a request is made to the /status endpoint.

Currently, the server module only supports GET requests to the /status
endpoint. The server module is added to the monitoring module and is
initialized when the monitoring module is initialized.

Currently the server module is not tested and is not configurable. The
server module is not configurable because the port is hardcoded to 3030
and the server module does not have any configuration options. This will
be addressed in a future commit.

Breaking Changes: None

Resolves: None