feat:refactor: Code improvements
Removes unwraps and replaces them with match statements to handle
errors. This will allow the program to continue running even if an
error occurs.

Fixes issue with output date format. Now the date is formatted as
YYYY-MM-DD HH:MM:SS.MI. This will make it easier to read the output
and compare the times.

Breaking changes: None

Resolves: None