feat: Improvement of the monitor configuration
Updates the config file to allow for more flexible configuration of the
monitor. Previously the monitors were configured with a list as the top
level object. This has been changed to the list to be one level deeper.

Breaking change: The configuration file now requires the monitors to be
configured under the "monitors" key.

Resolves: None