feat: Issue-1: Fix response problem with enums
Fixes a problem with the response when using enums in the response. The
problem is that the response is not correctly deserialized when using
enums when the response contains values.

The fix uses the serde attribute #[serde(tag = "status",
content = "information")] to correctly deserialize the response.

Breaking changes: Changes the response format for monitorstatus.

Resolves: Issue-1