feat: Add meminfo service
Adds a new service to read the /proc/meminfo file and return the data
as a JSON object. This service is used to monitor the memory usage of
the system. Only five fields are returned: MemTotal, MemFree,
MemAvailable, SwapTotal, and SwapFree. The data is returned at The
meminfo endpoint.

Breaking changes: None

Resolves: Issue-6