feat: Issue-16: Store PID file in /var/run
Based on the file hierarchy standard
<https://www.pathname.com/fhs/2.2/fhs-5.13.html>. The default place to
store the pid file should be /var/run. This commit changes the location
of the pid file to /var/run/monitoring-agent.pid.

Additionally, also fix an error in the testing of the server.

Breaking changes: The location of the pid file has changed from
/tmp/monitoring-agent.pid to /var/run/monitoring-agent.pid. This should
not affect any functionality of the monitoring agent.

Resolves: #16