feat:Issue-42:Store the monitoring in db
Stores the monitoring in the database. This will store it in a
mariadb/mysql database. The column names
- id - autoincrement
- log_time - always store as now(3)
- message - If error then include message.
- monitor_name - Monitor name
- status - Status of the monitor

If the configuration is set to store in the database then it will store
the monitoring in the database, else it will not. Currently, the
configuration allows for each monitor to decide if it should store in
the database or not, but this is not implemented yet.

Breaking changes: None

Resolves: #42