feat:Issue-56:Add loadaverage monitor
Adds a new monitor to the monitoring agent that will monitor the load
average of the system. The monitor will check the load average of the
system and compare it to the threshold set in the configuration file.
If the load average is above the threshold, the monitor will report an
error.

Also updates so that the check mthod is now required by by the Monitor.
This allows for the removal of multiple methods in the scheduler.

Breaking changes: None

Resolves: #56