feat:Issue-91: Fix tokio issue with postgres connections
Mysql/Mariadb driver does not use tokio, but postgres does so when
postgres queries it does a block_on. This causes the application to
crash as it blocks the main thread. To fix this make the database
service async. This will require a major rewrite as it will affect all
the code. To make the postgres connections use async we cannot use r2d2,
but must use bb8 connectionpool.

Breaking changes: None

Resolves: #91