feat:Issue-103: Update logging
Removes log4rs and replaces it with tracing logger. This is a breaking
change as the logging configuration file is no longer used. The logging
level is now via arguments. Logs both to sdout and a file and the
level can be different for each.

Adds instrumentation logging statements to the database service and the
procs so that any performance issues can be tracked.

Adds a few new logging statements to the monitors to help with
debugging.

Breaking changes: Logging file no longer used and new arguments for
setting logging level has been added.

Resolves: #103