feat: Issue-111: Implement cmdline reader
Implements the cmdline reader for reading the cmdline file in the proc
filesystem. The cmdline file contains the command line arguments passed
to the process. The reader reads the file and returns the command line
arguments, application name, and the process id.

Also fixes a few clippy issues.

Breaking changes: None

Resolves: #111