feat:Issue-218:Notification email
Adding email notification to the monitoring agent daemon. This will
allow the daemon to send an email to the user when an error is detected.

The system have configuration for resending an error after x minutes
and how often the system should check for new errors.

The url should have the followiing format.
- smtp://username:password@smtpserver:port?tls=required
tls is optional and can be set to require, optimistic or ignored.

Breaking changes: None

Resolves: #218