feat:Issue-214:Thresholds process monitor (#220)
Changes the process monitor to use threshold vales rather than only a
max value. This allows for more flexibility in the configuration of the
monitor. It has threshold for error and warn. If the process exceeds the
error threshold, the monitor will return an error. If the process exceeds
the warn threshold, the monitor will return a warning. If the process is
below the warn threshold, the monitor will return ok.

Breaking changes: max value is no longer used and will cause the agent
to fail to startup if it is used in the configuration.

Resolves: #214